requirements.txt
can
cantools
evdev
pygame
gpiozero
pillow
numpy
matplotlib
